requirements.txt
dash
dash-daq
plotly
pandas
pyserial
dash-extensions
paho-mqtt
